pyproject.toml
[project]
name = "sentinel-core"
version = "0.1.0"
description = "Autonomous Financial Analyst System using GraphRAG and Multi-LLM Providers"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Project Sentinel Team"}
]
keywords = ["finance", "ai", "graphrag", "llm", "django", "celery"]

dependencies = [
    # Web Framework - Django 6.0 with async support
    "django>=6.0",
    "channels>=4.0",
    "channels-redis>=4.2",
    "uvicorn[standard]>=0.30",
    "daphne>=4.1",
    
    # Task Queue - Celery with Redis
    "celery[redis]>=5.4",
    "redis>=5.0",
    
    # LLM Providers
    "groq>=0.9",
    "ollama>=0.3",
    "openai>=1.50",
    "anthropic>=0.30",
    "google-generativeai>=0.8",
    "together>=1.3",
    "httpx>=0.27",
    
    # Data Validation
    "pydantic>=2.9",
    "pydantic-settings>=2.4",
    
    # Graph Database - Neo4j
    "neo4j>=5.20",
    
    # LangChain for GraphRAG
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.2",
    
    # Serialization
    "msgpack>=1.0",
    "orjson>=3.10",
    
    # Tokenization
    "tiktoken>=0.7",
    
    # Async HTTP
    "aiohttp>=3.10",
    "aiofiles>=24.1",
    
    # Web Scraping
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    
    # Environment & Config
    "python-dotenv>=1.0",
    
    # Utilities
    "rich>=13.7",
    "tenacity>=8.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "black>=24.4",
    "pre-commit>=3.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src", "sentinel_core"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "sentinel_core.settings.development"
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
